+++ app/Http/Controllers/API/ReviewController.php
@@ -6,12 +6,13 @@ use App\Models\Review;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
 use Illuminate\Support\Facades\Validator;
+use Termwind\Components\Raw;
 
 class ReviewController extends Controller
 {
     public function index()
     {
-        $reviews = Review::all();
+        $reviews = Review::select('_id','book_id', 'rating', 'description')->get(); //Al ser publico, no quiero que me devuelva un user id 
         return response()->json([
             'status' => 'success',
             'data' => $reviews,
@@ -21,7 +22,6 @@ class ReviewController extends Controller
     public function store(Request $request)
     {
         $validator = Validator::make($request->all(), [
-            'user_id' => 'required|exists:users,id',
             'book_id' => 'required|exists:books,id',
             'description' => 'required|string',
             'rating' => 'required|integer|min:1|max:5',
@@ -33,8 +33,28 @@ class ReviewController extends Controller
                 'errors' => $validator->errors(),
             ], 422);
         }
+        // Obtener el usuario autenticado desde el middleware AuthRole
+        $user = $request->user();
+        // Verificar si el usuario ya ha reseñado este libro
+        $existingReview = Review::where('user_id', $user->id)
+            ->where('book_id', $request->book_id)
+            ->first();
 
-        $review = Review::create($request->all());
+        if ($existingReview) {
+            return response()->json([
+                'status' => 'error',
+                'message' => 'You have already reviewed this book',
+            ], 403);
+        }
+
+        // Crear la reseña con el user_id del usuario autenticado, ya que el usuario es el unico que puede crear reseñas en su nombre
+        $review = Review::create([
+            'user_id' => $user->id,
+            'book_id' => $request->book_id,
+            'description' => $request->description,
+            'rating' => $request->rating,
+        ]);
+
         return response()->json([
             'status' => 'success',
             'message' => 'Review created successfully',
@@ -44,7 +64,7 @@ class ReviewController extends Controller
 
     public function show($id)
     {
-        $review = Review::find($id);
+        $review = Review::where('_id', $id)->select('_id', 'book_id', 'description', 'rating')->first();;
         if (!$review) {
             return response()->json([
                 'status' => 'error',
@@ -60,7 +80,7 @@ class ReviewController extends Controller
 
     public function update(Request $request, $id)
     {
-        $review = Review::find($id);
+        $review = Review::where('_id', $id)->first();
         if (!$review) {
             return response()->json([
                 'status' => 'error',
@@ -68,8 +88,18 @@ class ReviewController extends Controller
             ], 404);
         }
 
+        // Obtener el usuario autenticado,  ya que este podra modificar unicamente las reseñas hechas por el
+        $user = $request->user();
+
+        // Verificar si el usuario es el creador de la reseña
+        if ($review->user_id !== $user->id) {
+            return response()->json([
+                'status' => 'error',
+                'message' => 'You are not authorized to update this review',
+            ], 403);
+        }
+
         $validator = Validator::make($request->all(), [
-            'user_id' => 'sometimes|required|exists:users,id',
             'book_id' => 'sometimes|required|exists:books,id',
             'description' => 'sometimes|required|string',
             'rating' => 'sometimes|required|integer|min:1|max:5',
@@ -81,8 +111,8 @@ class ReviewController extends Controller
                 'errors' => $validator->errors(),
             ], 422);
         }
-
-        $review->update($request->all());
+        //El user_id no es modificable
+        $review->update($request->only(['book_id', 'description', 'rating']));
         return response()->json([
             'status' => 'success',
             'message' => 'Review updated successfully',
@@ -90,14 +120,24 @@ class ReviewController extends Controller
         ], 200);
     }
 
-    public function destroy($id)
+    public function destroy( Request $request, $id)
     {
-        $review = Review::find($id);
+        $review = Review::where('_id', $id)->first();
         if (!$review) {
             return response()->json([
                 'status' => 'error',
                 'message' => 'Review not found',
             ], 404);
+        }
+        // Obtener el usuario autenticado
+        $user = $request->user();
+
+        // Verificar si el usuario es el creador de la reseña o un admin, ya que estos son los unicos que pueden eliminarla
+        if ($review->user_id !== $user->id && $user->role !== 'admin') {
+            return response()->json([
+                'status' => 'error',
+                'message' => 'You are not authorized to delete this review',
+            ], 403);
         }
 
         $review->delete();
@@ -105,5 +145,6 @@ class ReviewController extends Controller
             'status' => 'success',
             'message' => 'Review deleted successfully',
         ], 204);
+        
     }
 }
